feat(core): widen length path to 32-bit, lifting the 64KB SDU cap
The reassembly buffer, transport, and request/response dispatch were
uint16_t end-to-end, capping a single ISO-TP SDU at 65535 bytes. With
the escape FirstFrame now carrying a 32-bit FF_DL, widen the whole
length path to uint32_t so large CAN-FD transfers work end to end:

- transport: msg_len, bytes_processed, tx_sdu_size/len, send length
- core: uds_input_sdu, handle_request/execute_handler, send_response,
  client_request, parse_addr_len, and the truncating client-cb cast
- config: rx_buffer_size, tx_buffer_size, and the length-bearing
  callbacks (tp_send, service handler, did_write, response_cb, is_safe)

BREAKING: integrators implementing the widened callbacks must change
their length parameter from uint16_t to uint32_t. DID/security buffer
capacities and resp_len are unchanged. In-repo examples and tests
updated to match.

Tested: new >64KB escape-FF TX/RX reassembly cases (70000-byte SDU)
plus the existing escape/FC suite; full suite 45/45 green.

+++ examples/freertos_demo/main.c
@@ -37,7 +37,7 @@ static uds_config_t cfg;
 static uint8_t rx_buf[1024];
 static uint8_t tx_buf[1024];
 
-static int mock_send(struct uds_ctx *ctx, const uint8_t *data, uint16_t len)
+static int mock_send(struct uds_ctx *ctx, const uint8_t *data, uint32_t len)
 {
     return 0;
 }
